Move Road construction to Level class
