+++ lib/game.js
@@ -1,6 +1,7 @@
 const Toad = require('./toad.js');
 const Home = require('./home.js');
 const Score = require('./score.js')
+const Level = require('./level.js')
 const {Platform, Turtle, Crocodile} = require('./platform.js');
 const {Obstacle, Auto, Tractor, SlowCar, Semi, FastCar, Snake} = require('./obstacle.js');
 
@@ -13,20 +14,14 @@ class Game {
     this.turtles = [];
     this.logs = [];
     this.autos = [];
-    this.tractors = [];
-    this.slowCars = [];
-    this.fastCars = [];
-    this.semis = [];
+    this.level = new Level(this.squareWidth);
     this.homes = this.buildHomes();
     this.score = new Score();
     this.toad = new Toad(475, 662.5, 'safe', this.homes, this.score)
   }
 
   buildLevel() {
-    this.autos = this.buildRoadRow(Tractor, 614.5, this.squareWidth, -2, this.squareWidth * 4, 4, this.squareWidth * 4, 'yellow');
-    this.autos = this.autos.concat(this.buildRoadRow(SlowCar, 564.5, this.squareWidth * 2, 1, this.squareWidth * 5, 3, this.squareWidth * 4, 'pink'))
-    this.autos = this.autos.concat(this.buildRoadRow(FastCar, 514.5, this.squareWidth * 2, 3, this.squareWidth * 3, 4, this.squareWidth * 3, 'red'));
-    this.autos = this.autos.concat(this.buildRoadRow(Semi, 464.5, this.squareWidth * 5, -3, this.squareWidth * 3, 3, this.squareWidth * 4, 'orange'));
+    this.autos = this.level.buildRoadLevel();
     this.turtles = this.buildRiverRow(Turtle, 364.5, this.squareWidth * 3, -2, this.squareWidth * 3, 6, this.squareWidth * 0);
     this.logs = this.buildRiverRow (Platform, 314.5, this.squareWidth * 3, 3, this.squareWidth * 2, 4, this.squareWidth * 2);
     this.logs = this.logs.concat(this.buildRiverRow (Platform, 264.5, this.squareWidth * 7, 4, this.squareWidth * 3, 3, this.squareWidth * 8));
@@ -65,22 +60,6 @@ class Game {
     this.context.fillRect(0, 0, this.canvas.width, 412.5);
 
 
-  }
-
-  buildRoadRow(vehicleType, yCoordinate, carLength, speed, gap, numCars, offset, color) {
-    let Type = vehicleType;
-    let vehicleyCoordinate = yCoordinate;
-    let vehicleLength = carLength;
-    let vehicleSpeed = speed;
-    let vehicleGap = gap;
-    let numVehicles = numCars;
-    let rowOffset = offset;
-    let roadArray = [];
-
-    for (let i = 0; i < numVehicles; i++) {
-      roadArray.push(new Type((vehicleLength + vehicleGap) * i, vehicleyCoordinate, vehicleLength, vehicleSpeed, vehicleGap, rowOffset, color));
-    }
-    return roadArray;
   }
 
   buildRiverRow (PlatformType, yCoordinate, platformLength, platformSpeed, platformGap, numPlatforms, platformOffset) {
